Change types to not assume an arch
Some of these types were assuming that `size_t == uint64_t` or that
`uint64_t == unsigned long` but that's not universally true, some BSDs
such as Darwin define uint64_t as unsigned long long which warns on
literals that are defined as `UL`.

Additionally there were also some assumptions made on the pch as to
forward declarations. So I explicitly added some needed ones.

+++ source/common/memory/stats.cc
@@ -7,13 +7,13 @@
 namespace Memory {
 
 uint64_t Stats::totalCurrentlyAllocated() {
-  uint64_t value = 0;
+  size_t value = 0;
   MallocExtension::instance()->GetNumericProperty("generic.current_allocated_bytes", &value);
   return value;
 }
 
 uint64_t Stats::totalCurrentlyReserved() {
-  uint64_t value = 0;
+  size_t value = 0;
   MallocExtension::instance()->GetNumericProperty("generic.heap_size", &value);
   return value;
 }
